# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIPrintVerse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List stores (admin only)
      responses:
        "200":
          description: Stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create store (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{id}:
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Delete store (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # Analytics
...
  /seo/auto-link-keywords:
    get:
      operationId: listAutoLinkKeywords
      tags: [seo]
      summary: List auto-link keywords
      responses:
        "200":
          description: Keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutoLinkKeyword"
    post:
      operationId: createAutoLinkKeyword
      tags: [seo]
      summary: Create auto-link keyword (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoLinkKeywordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoLinkKeyword"
  /seo/auto-link-keywords/{id}:
    delete:
      operationId: deleteAutoLinkKeyword
      tags: [seo]
      summary: Delete auto-link keyword (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /seo/article-templates:
    get:
      operationId: listArticleTemplates
      tags: [seo]
      summary: List article templates (admin only)
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArticleTemplate"
    post:
      operationId: createArticleTemplate
      tags: [seo]
      summary: Create article template (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleTemplate"
  /seo/generation-batches:
    get:
      operationId: listGenerationBatches
      tags: [seo]
      summary: List generation batches (admin only)
      responses:
        "200":
          description: Batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GenerationBatch"
    post:
      operationId: createGenerationBatch
      tags: [seo]
      summary: Create generation batch (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerationBatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationBatch"

  # AI features
...
  /ebooks:
    get:
      operationId: listEbooks
      tags: [ebooks]
      summary: List ebooks
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Ebooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ebook"
    post:
      operationId: createEbook
      tags: [ebooks]
      summary: Create a new ebook (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EbookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ebook"
  /ebooks/{id}:
    get:
      operationId: getEbook
      tags: [ebooks]
      summary: Get ebook by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ebook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EbookWithChapters"
        "404":
          description: Not found
    patch:
      operationId: updateEbook
      tags: [ebooks]
      summary: Update ebook (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EbookUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ebook"
    delete:
      operationId: deleteEbook
      tags: [ebooks]
      summary: Delete ebook (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /ebooks/{id}/chapters:
    get:
      operationId: listEbookChapters
      tags: [ebooks]
      summary: List chapters for an ebook
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EbookChapter"
    post:
      operationId: addEbookChapter
      tags: [ebooks]
      summary: Add chapter to ebook (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EbookChapterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EbookChapter"
  /ebooks/{id}/generate-from-posts:
    post:
      operationId: generateEbookFromPosts
      tags: [ebooks]
      summary: Auto-generate ebook chapters from blog posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EbookFromPostsInput"
      responses:
        "200":
          description: Generated chapters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EbookWithChapters"
  /ebooks/{id}/export-pdf:
    post:
      operationId: exportEbookPdf
      tags: [ebooks]
      summary: Export ebook as PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EbookExportResult"
...
components:
  schemas:
...
    Store:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        store_url:
          type: string
        platform:
          type: string
        username:
          type: ["string", "null"]
        user_id:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, name, store_url, platform, user_id, created_at, updated_at]
    StoreInput:
      type: object
      properties:
        name:
          type: string
        store_url:
          type: string
        platform:
          type: string
        username:
          type: string
        user_id:
          type: string
      required: [name, store_url, platform, user_id]

    # Analytics
...
    AutoLinkKeyword:
      type: object
      properties:
        id:
          type: string
        keyword:
          type: string
        target_post_id:
          type: ["string", "null"]
        priority:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
      required: [id, keyword, priority, is_active, created_at]
    AutoLinkKeywordInput:
      type: object
      properties:
        keyword:
          type: string
        target_post_id:
          type: string
        priority:
          type: integer
        is_active:
          type: boolean
      required: [keyword]

    # SEO Templates
    ArticleTemplate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        template_type:
          type: string
        title_template:
          type: string
        slug_template:
          type: string
        content_template:
          type: string
        excerpt_template:
          type: ["string", "null"]
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_at:
          type: string
      required: [id, name, template_type, title_template, slug_template, content_template, category, is_active, created_at]
    ArticleTemplateInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        template_type:
          type: string
        title_template:
          type: string
        slug_template:
          type: string
        content_template:
          type: string
        excerpt_template:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        is_active:
          type: boolean
      required: [name, title_template, slug_template, content_template]
    GenerationBatch:
      type: object
      properties:
        id:
          type: string
        batch_name:
          type: string
        template_id:
          type: ["string", "null"]
        total_articles:
          type: integer
        generated_count:
          type: integer
        published_count:
          type: integer
        status:
          type: string
        created_at:
          type: string
      required: [id, batch_name, total_articles, generated_count, published_count, status, created_at]
    GenerationBatchInput:
      type: object
      properties:
        batch_name:
          type: string
        template_id:
          type: string
        total_articles:
          type: integer
        variables_data:
          type: object
        created_by:
          type: string
      required: [batch_name]

    # AI
...
    Ebook:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        subtitle:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        author_name:
          type: string
        cover_image:
          type: ["string", "null"]
        language:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        pdf_url:
          type: ["string", "null"]
        epub_url:
          type: ["string", "null"]
        isbn:
          type: ["string", "null"]
        page_count:
          type: ["integer", "null"]
        word_count:
          type: ["integer", "null"]
        meta_title:
          type: ["string", "null"]
        meta_description:
          type: ["string", "null"]
        print_ready:
          type: boolean
        published_at:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, title, slug, author_name, language, category, status, print_ready, created_at, updated_at]
    EbookInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        subtitle:
          type: string
        description:
          type: string
        author_name:
          type: string
        cover_image:
          type: string
        language:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        meta_title:
          type: string
        meta_description:
          type: string
      required: [title, slug]
    EbookUpdate:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        description:
          type: string
        author_name:
          type: string
        cover_image:
          type: string
        language:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        pdf_url:
          type: string
        epub_url:
          type: string
        isbn:
          type: string
        print_ready:
          type: boolean
        meta_title:
          type: string
        meta_description:
          type: string
        published_at:
          type: string
    EbookChapter:
      type: object
      properties:
        id:
          type: string
        ebook_id:
          type: string
        blog_post_id:
          type: ["string", "null"]
        title:
          type: string
        order_index:
          type: integer
        content:
          type: ["string", "null"]
        is_generated:
          type: boolean
        created_at:
          type: string
      required: [id, ebook_id, title, order_index, is_generated, created_at]
    EbookChapterInput:
      type: object
      properties:
        blog_post_id:
          type: string
        title:
          type: string
        order_index:
          type: integer
        content:
          type: string
        is_generated:
          type: boolean
      required: [title]
    EbookWithChapters:
      type: object
      properties:
        ebook:
          $ref: "#/components/schemas/Ebook"
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/EbookChapter"
      required: [ebook, chapters]
    EbookFromPostsInput:
      type: object
      properties:
        postIds:
          type: array
          items:
            type: string
        category:
          type: string
        autoOrder:
          type: boolean
      required: [postIds]
    EbookExportResult:
      type: object
      properties:
        pdf_url:
          type: ["string", "null"]
        epub_url:
          type: ["string", "null"]
        message:
          type: string
        status:
          type: string
      required: [message, status]
...