# lib/api-zod/src/generated/types/articleTemplateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIPrintVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleTemplateInput {
  name: string;
  description?: string;
  template_type?: string;
  title_template: string;
  slug_template: string;
  content_template: string;
  excerpt_template?: string;
  category?: string;
  tags?: string[];
  is_active?: boolean;
}