# lib/api-zod/src/generated/types/ebookChapterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIPrintVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EbookChapterInput {
  blog_post_id?: string;
  title: string;
  order_index?: number;
  content?: string;
  is_generated?: boolean;
}