# lib/api-zod/src/generated/types/ebookUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIPrintVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EbookUpdate {
  title?: string;
  subtitle?: string;
  description?: string;
  author_name?: string;
  cover_image?: string;
  language?: string;
  category?: string;
  tags?: string[];
  status?: string;
  pdf_url?: string;
  epub_url?: string;
  isbn?: string;
  print_ready?: boolean;
  meta_title?: string;
  meta_description?: string;
  published_at?: string;
}