# lib/api-zod/src/generated/types/generationBatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIPrintVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerationBatch {
  id: string;
  batch_name: string;
  /** @nullable */
  template_id?: string | null;
  total_articles: number;
  generated_count: number;
  published_count: number;
  status: string;
  created_at: string;
}